Refactor CRUD services hierarchy and separate classic CRUD from weather data processing
