+++ src/main/java/ru/bukhtaev/service/WeatherProcessingServiceImpl.java
@@ -2,26 +2,39 @@ package ru.bukhtaev.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.validation.annotation.Validated;
+import org.springframework.transaction.annotation.Transactional;
+import ru.bukhtaev.exception.DataNotFoundException;
 import ru.bukhtaev.model.Weather;
+import ru.bukhtaev.repository.jpa.IWeatherJpaRepository;
 import ru.bukhtaev.validation.MessageProvider;
 
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
+import static org.springframework.transaction.annotation.Isolation.READ_COMMITTED;
 import static ru.bukhtaev.util.Utils.round;
-import static ru.bukhtaev.validation.MessageUtils.MESSAGE_CODE_FAILED_TO_COMPUTE;
-import static ru.bukhtaev.validation.MessageUtils.MESSAGE_CODE_THERE_IS_NO_DATA;
+import static ru.bukhtaev.validation.MessageUtils.*;
 
 /**
  * Реализация сервиса обработки данных о погоде.
  */
 @Service
-@Validated
+@Transactional(
+        isolation = READ_COMMITTED,
+        readOnly = true
+)
 public class WeatherProcessingServiceImpl implements IWeatherProcessingService {
+
+    /**
+     * Репозиторий данных о погоде.
+     */
+    private final IWeatherJpaRepository weatherRepository;
 
     /**
      * Сервис предоставления сообщений.
@@ -31,11 +44,51 @@ public class WeatherProcessingServiceImpl implements IWeatherProcessingService {
     /**
      * Конструктор.
      *
-     * @param messageProvider сервис предоставления сообщений
+     * @param weatherRepository репозиторий данных о погоде
+     * @param messageProvider   сервис предоставления сообщений
      */
     @Autowired
-    public WeatherProcessingServiceImpl(final MessageProvider messageProvider) {
+    public WeatherProcessingServiceImpl(
+            final IWeatherJpaRepository weatherRepository,
+            final MessageProvider messageProvider
+    ) {
+        this.weatherRepository = weatherRepository;
         this.messageProvider = messageProvider;
+    }
+
+    @Override
+    public List<Weather> getTemperatures(final String cityName) {
+        return weatherRepository.findAll()
+                .stream()
+                .filter(weather -> weather.getCity().getName().equals(cityName)
+                        && weather.getDateTime().toLocalDate().equals(LocalDate.now())
+                )
+                .toList();
+    }
+
+    @Override
+    public Double getTemperature(final String cityName, final ChronoUnit timeUnit) {
+        final LocalDateTime now = LocalDateTime.now();
+
+        final Weather weather = weatherRepository.findAllByCityName(cityName)
+                .stream()
+                .filter(w -> w.getDateTime().truncatedTo(timeUnit)
+                        .equals(now.truncatedTo(timeUnit)))
+                .findFirst()
+                .orElseThrow(() -> new DataNotFoundException(
+                        messageProvider.getMessage(
+                                MESSAGE_CODE_TEMPERATURE_NOT_FOUND,
+                                cityName
+                        )
+                ));
+
+        return weather.getTemperature();
+    }
+
+    @Override
+    @Transactional(isolation = READ_COMMITTED)
+    public void delete(final String cityName) {
+        weatherRepository.deleteAllByCityName(cityName);
     }
 
     @Override
